updates check on allowed component combinations
Provides a more flexible initialisation of the Model class (hopefully future-proof, i.e. addition of components, e.g. marine or groundwater) by checking for all possible combinations of components that can be given (based on a full factorial design of experiment).

A component given to Model can be a subclass of Component, a DataBase, or None. It checks for that.

It checks if DataBase is only given when required (i.e. if not all components downwards are only DataBase or None), otherwise it returns a UserWarning.
It checks if None is only given when not required (i.e. if all components downwards are None).

Note, the case (surface=None, subsurface=None, openwater=None) does not raise any exception at the moment. Yet to decide if this is an acceptable setup or not.

The check for compatibility between Components inputs and outputs (i.e. genuine components SurfaceComponent, SubSurfaceComponent, or OpenWater) is not done during instantiation of Model because the function prototype of their 'run' method is not known until runtime, so these checks still happen in the __call__ method of Component.