redesigns time_ around cf.Field and cfunits.Units
in view to work with cf-compliant datasets, this commit renames the TimeFrame to TimeDomain, and it redesigns the class around cf.Field and cfunits.Units.

The new TimeDomain inherits from cf.Field to be able to set constructs to its domain, that can later be used to compare this instance of Field with e.g. a field in a CF-netCDF file. It also uses cfunits.Units to deal with reference time (like 'seconds since 1970-01-01 00:00:00' e.g.).

Unlike TimeFrame, the new class TimeDomain is not as restrictive on the epoch time used. But if TimeDomain is instantiated from a list of datetime objects, the UNIX epoch is used arbitrarily as the epoch time. reference.