replaces check on domain by check on specific construct for TimeDomain ==
Because when comparing an actual Field from a dataset to this "artificial" field for time considerations, a domain.equals would fail because of the other dimensions of the dataset.

An alternative solution would be to merge the TimeDomain and SpaceDomain before comparing domains with the actual Field dataset, but this would not be very convenient, as a third Field from the two Fields TimeDomain and SpaceDomain would need to be created just for this purpose. Moreover, there may be some constructs in the domain of the actual Field dataset that don't need to be checked.