redesigns space_ around cf.Field
in view to work with cf-compliant datasets, this commit provides a first proper attempt at implementing a SpaceDomain, for a Grid in this case.

The Grid class can work with proper lat/lon grids, or rotated grids for now. It is a subclass of cf.Field with constructs set for the space domain constructs, including lat/lon (or grid_lat/grid_lon), altitude (optional), and the coordinate reference in case of a rotated grid (coordinate reference providing the information necessary to convert grid_lat/grid_lon to lat_lon)

The method to compare Fields on their space domain constructs will only compare their lat/lon (or grid_lat/grid_lon), altitude (optional), and the coordinate_conversion and datum of the coordinate_reference construct. It does not compare the whole of the coordinate_reference directly using its .equals() method because this also includes a check on the size of the collection of dimensions of each Field. This is not desirable to do this check as the time dimension is likely to be present in the Field but not in the SpaceDomain (for obvious reasons), and this also extends to the altitude dimension, which may be of size 1 in the Field, so that it allows to avoid defining it when instantiating the Grid.