moves check on data/parameter + adds check on domain
The existing checks on data and parameter being available / provided implemented at the Component level are moved to the Model level to avoid checking several times if Component is called several times (e.g. in case of parallelisation or else). The checks on inwards stays ay the component level for two reasons: first because the check on outwards is at this level too for now, and second because it is likely to be dealt with metaclasses in the future so these checks should become irrelevant.

This commit also adds a check on the time and space domain compatibilities between the data and the components. It checks for time and space for driving data, and for space only for ancillary data.

The API of Model.simulate is also slightly modified and the concept of "context" disappears and is replaced by "domain" (TimeDomian, SpaceDomain) and "data" (DataBase) being provided separately.