replaces NoneComponent by a NullComponent (philoshophy shift)
Previously, a NoneComponent could be used in place of a genuine Component (e.g. SurfaceLayerComponent), and this component was returning an empty dictionary (i.e. used to completely turn off a component in the model). Given that the interface between the components was only one-way coupled, there were some situations where this was possible (e.g. OpenWaterComponent, only outputting discharge to the ocean, which is a spatial boundary of the model, could be turned off because not other component was relying on it). This was not satisfactory.

Instead, a new NullComponent is created. Unlike the NoneComponent, it returns zeros for its outwards "fluxes", meaning that if e.g. a component rely on OpenWaterComponent in the future (e.g. inundation for the SurfaceLayerModel) i.e. two-way coupled, the interface is still satisfied, the fluxes are simply null. By setting the component to NullComponent, the user is explicitly stating that the outwards "fluxes" of the component are negligible/ignored. Before None should be given when instantiating a Model, now the class NullComponent must be given.

Moreover, previously using data in place of a component was possible by giving a DataBase when instantiating a Model. This was problematic because checks on the availability of the outwards "fluxes" (and there spatio-temporal relevance) needed to be handled as a special case in comparison with genuine Components. Now, the class DataComponent must be given in place of the DataBase, and the DataBase needs to be given when calling the "simulate" method of Model. By artificially setting the outwards "fluxes" of the component it replaces, the availability and the space/time compatibility of the data given can now be checked the same way forcing data are checked. DataComponent will then behave as a simple tunnel, taking the data for the "fluxes" as input and returning them as output without any processing on it at all.

These changes simplify considerably the type checks that were done in the init of Model. Now it only checks if a genuine Component is given, that it is the relevant type (e.g. SurfaceLayerComponent for the surfacelayer category). If a DataComponent/NullComponent is given, it is initialised using the information on the interface (especially outwards) of the component it replaces.

The unittest on the API needed to be modified accordingly.

The API itself is also changed, breaking backwards compatibility, since only subclasses of _Component can be given for surfacelayer/subsurface/openwater during the instantiation of Model (not a DataBase/None anymore).