makes framework unit-aware
Up until now, the framework was not checking for the units of the data given to each component. This is now the case by relying on cfunits. One of the advantages of using cfunits is that it can understand that e.g. "kg m-2 s-1" is the same as "kg/m2/s".

The way a class for a _Component should be defined has changed as a consequence, and now the driving/ancil/parameters should be given as a dictionary (key: name, value: unit) in place of the tuple of names previously required. At the moment, there is no check on the units outputted by each _Component, as this is meant to be a set characteristic of the framework, and expect the modeller to produce a component that outputs in the right (probably SI) unit.