adds distinction between category of class, and category of instance
The private class attribute "_cat" is used to set the category of an instance of the class. But you want to use something else to set the category when a DataComponent or a NullComponent is used, because it is useful to know what the component is substituting for (i.e. SurfaceLayerComponent, SubSurfaceComponent, orOpenWaterComponent). So _cat needed to be renamed to _kind, to avoid the confusion, as they need to be different in some situations.

_kind is used to set the category when instantiating a genuine component, while the new classmethod get_kind() is used on the substituting_class for DataComponent and NullComponent.