adds silent instantiations of objects for data and parameter arg in Model.simulate()
Because some components do not require data (in particular NullComponent, but not only), or do not require parameters (less likely, but possible for very simple models), the parameters and data arguments for the 3 components because keyword arguments with default value set to None (i.e. optional).

If data is set to None by user, or not set at all by user, a dummy instance of DataBase is generated, and it goes through the _check_component_data that will deal with the actual relevance of not giving any data given the context.

If parameter is set to None by user, or not set at all by user, an empty dictionary is generated, and it goes through _check_component_parameters that will deal with the actual relevance of not giving any parameter given the context.