adds draft of loop over time & adds initialise+finalise methods
A first loop over time is added to the framework at the Model.simulate() level, and the data is read in at the _Component.__call__() level. It is currently using arbitrarily the time domain of the SurfaceLayerComponent, because the framework only allows for the 3 components to have identical time domains.

This commit is also adding overdue `initialise` and `finalise` methods to be implemented alongside the `run` method in the user-defined model components (e.g. Dummy components) To enforce the defintion of these methods, metaclass abc.ABCMeta is used for _Component, SurfaceLayerComponent, SubSurfaceComponent, and OpenWaterComponent. While, DataComponent and NullComponent (being used in replacement of user-defined components) implement their own `initialise` and `finalise` methods. However, this still does not enforce the framework interface between components (i.e. the inwards and outwards).