adds 'states_info' in the defintion of a Component
Because state values need to be exchanged between modelling time steps, they need to be part of the definition of a Component, alongside driving/ancillaries/parameters.

The `initialise` method is expected to yield intitial values for the component state values, and the `finalise` method is expected to store the final component state values (if continuation of the simulation is required).