renames DataBase to Dataset to be more compliant with unidata data model
Also DataBase was probably a stretch of what the object is and does, i.e. a mapping of names to Variables. It has no SQL query capability or anything of such sort, and it is unlikely to require any in the future.

Uses DataSet rather than Dataset to avoid naming conflict in namespace with netcdf4.Dataset.