relaxes the comparison of TimeDomain with Variable
Previously, the `is_matched_in` method was checking that the two 'time' constructs were equal by using the `equals` method of a cf.Construct. However, if a date-time data was defined with different reference times, but representing the same day, it was returning False.

The new `is_time_equal_to` method of TimeDomain is now less strict on the comparison, and allows for instances to have different reference times, as long as they have the same calendar (because cf.Units cannot compare datetimes in different calendars).