corrects error in indexing for Clock
The Clock was updating the current_timeindex and the current_datetime after indexing the switches, which means that it was assigning the next current_timeindex and current_datetime attributes, not the one matching the switches. So it needs to do the increment before indexing the switches, so that the switches and the current timeindex/datetime are actually relating to the same point in time. Beyond being incorrect, it should have resulted in an out of bounds exception when trying to index the datasets in time, but because the Clock is stopping one datetime before the end (because the end_datetime is the end of the last timestep, not its start), and the dummy dataset which was used for testing had a value for the end_datetime regardless, hence there was no exception raised.

The correction includes:
- fixing the end_timeindex which is meant to be the supermesh length minus one (because indexing of array starts from 0...)
- initialising the current_timeindex and current_datetime to one date prior the actual start of the supermesh, because the iterator needs to increment before indexing the switches so that attributes and returned values are for the same point in time
- moving the time increment before indexing the switches in `__next__` method