adds leading and trailing truncation options to compare TimeDomains
Because driving datasets are not always required for the last TimeDomain datetime (because the end datetime is the end of the last timestep, not its start), this adds optional private arguments to the `is_time_equal_to` method of `TimeDomain` to allow for the use of a trailing index (negative integer expected) to ignore the last (if -1 is used) datetime of TimeDomain when comparing it with e.g. a Variable.

Adds a leading index option too (positive integer expected), in case the first (or more) datetime are not to be compared.