adds current datetime as an accessible argument for Component.run method
If the user requires in their bespoke component to know e.g. the day of the month or the month of the year, they can now add a `datetime` argument in the `run` method prototype and they will get access to a datetime.datetime instance for the start of the current timestep.