uses `shape_` attribute of SpaceDomain in Dummy components and NullComponent
In order for the NullComponent to generate as an output an array of the relevant dimension (i.e. the one set by the user), a new atrribute `shape_` is given to SpaceDomain to know the dimension expected. It is a tuple with the order (Z,) Y, X (Z if altitude is not ignored).

This new shape_ attribute is then passed by Model to Component as a keyword argument `spaceshape`, and as a result is accessible to NullComponent, and to any user-defined component for that matter.

As a matter of fact Dummy components are now using spaceshape to generate outputs of the relevant shape. In practice, the inputs would be used to infer the outputs one way or another so that outputs should necessarily inherit the shape of the input.