updates docstring checking methods of Model to stop using private _Component
As its name is hinting _Component is a "private" class of the components module, and should not be used outside of its module.

Since _Component was only necessary in the docstrings for the methods checking on domain/paramaters/data of a given component, it was replaced by a Union[] listing all the possible component classes that could be tested (SurfaceLaterComponent, SubSurfaceComponent, OpenWaterComponent, DataComponent, NullComponent).