merges surface and subsurface runoffs into simply runoff
There is no need to distinguish surface and subsurface runoff through the interface, as they are both inputs to the river channel, so they can be aggregated into one flux.

This was actually already the case in the draft framework diagram for the project, it was a misinterpretation of it from me.