adds constants to the definition of a Component
Constants are distinct from parameters in the sense that they do not require tuning. Constants can be physical measures (e.g. the gravitational acceleration, liquid water density, etc.), but also related to the structure of the component itself (e.g. number of soil layers, number of plan functional types, etc.). This second type of constant is expected to be required for the initialisation of a model, and not only to the simulation itself.

Hence constants are provided both to the 'initialise' and 'run' methods of the Component. Like parameters, constants are optional when calling the 'simulate' method of Model (because some components may feature neither of these). Unlike parameters, constants are expected to be given default values in the definition of 'run' method of the Component, so that no check is made on their availability in the arguments of 'simulate'.

Dummy components are updated accordingly.