adds solver_history to the Components
In order to implement an interface, each component definition must specify the number of previous timesteps to hold in memory for their solving scheme (multi-step numerical solving methods). This information is expected in the __init__ of the given Component, and will be held as an attribute for the interface to have access to it (when implemented).

The Component's initialise method must intialise the right number of timesteps for its states accordingly.

DataComponent and NullComponent are given a value of 0 for their solver_history because they are not actual models, so they do not need model states.

Dummy components are impacted accordingly.