adds first implementation of an interface with functionalities
Up until now, the interface was simply a dictionary holding/mapping any variable it was given.

An Interface class (in an interface module) is now implemented. It inherits from avbc.MutableMapping, and as such, it behaves exactly like a dictionary. However, it also features some metadata (i.e. information about which variables are states and which variables are fluxes). It also features an increment_states method to be able to roll the state values (t becomes t-1, t-1 becomes t-2, etc. and t reset to zero).

For a given component state, the decision was made to distinguish its different timesteps with a naming "trick": state for t is named "state", state for t-1 is named "state_", state for t-2 is named "state__", etc. (adding one more trailing underscore each time). This decision was made for two reasons:
- if previous timesteps are stored as an additional dimension in the variable array, there may be confusing as to which axis it is (although arguably it would be the first axis), and also which order of this axis is a travel back in time (i.e. index 0 for current timestep or index -1?)
- if the alternative using "big" numpy array is chosen, the only easy possibility to increment the states would be to use numpy.roll function, but it does not feature an inplace option, so that a new array would need to be created (which is not the case with the implement proposed here)

/!\ This implementation is only temporary because it will not work with components running for different temporal resolutions. The reason being that Interface does not distinguish states according to the component they belong to, and as such, it can only be called once for each loop in the Clock. But not necessarily all components are switched on for this iteration if they are not all on the same temporal resolution.
However, it is tested and works for Dummy components with the same temporal resolution.