implements differentiation in states increment + replaces trailing underscores by negative index
As mentioned in commit 502cafaef586953933ef0b7504ebd47b8bea8a7e the interface as implemented was not working for components with different temporal resolutions because it features only one increment method. This is now fixed, and it features 3 methods, one for each component kind.
+
In addition, the decision previously made to identify the different timesteps for a state with trainling underscores was not very practical if a lot of timesteps were to be stored, and it was not very clean approach. Instead, a new class State was created, it stores of the timesteps for a given component state in a list, and it features a custom indexing for this list as described in its docstring.

The Dummy components are updated to reflect these changes.