moves States from Interface to Component
The main reason for this move is that with the current implementation, the Components cannot own states with the same name. Even if this is unlikely, this is not impossible, and because they are stored in a dictionary, they are silently overwritten by the Component's methods. This also makes the code cleaner, because there is no need for component specific methods, as was the case in Interface, because _Component knows what kind it is.

The State class is moved to the components module, to be used by _Component. The initialise and increment methods for the states are now methods of _Components.

The Interface is now only responsible for holding the fluxes now. The Interface object is a bit redundant with just using a dictionary directly, but Interface will be responsible for regridding and supermesh considerations in the future, so keep it as its own class.