moves timedomain/spacedomain/data/parameters/constants to Component.__init__() rather than Model.simulate()
In the previous implementation of Model, a lot of actions were repeated because they had to be done for each Component individually, which really meant that these actions should have belonged to the Component itself. This was designed this way because a Model could be reused with different resolutions/data/parameters/etc without having to create a new instance of Model, but this made for a rather messy simulate method of Model, and also for a long list of argument for its prototype.

Before, a Model would be init with a class for each Component, now it needs an instance for each Component. Most of the checks that were done in Model are now done in _Component.

This has the effect of slightly complexifying the blueprint of a user-defined component (i.e. like the Dummy ones), but this is more logical for a Component to get as arguments (and hold as attributes these attributes) timedomain/spacedomain/parameters/constants information.

Also, spaceshape and timestepinseconds are now attributes of _Component, and as such initialise/run/finalise methods have access to it directly through self, no need to pass them in kwargs anymore.

Updated the dummy Components accordingly, as well as the unittests in test_model (which also meant including some dummy_data in the tests directory).