fixes timeindex given to each Component
Because Components can run on different temporal resolutions, the timeindex that was given to the Components during the loop of Clock was the index corresponding to the temporal supermesh (i.e. the fastest running Component). The problem was that this index is then used to get the right driving data when calling the Component. This did not raise any error just yet, because there is a NotImplementedError is Model if the timedomains are not equal, but this will not always be the case.

Now the get_current_timeindex method of Clock takes an argument 'component_type', and depending on its value, it returns the correct index where the relevant timestep should be found in the dataset for that Component.

Unittests are modified to test this new feature.