gets rid of Variable class
The Variable class was a subclass of cf.Field with one custom method to create a Variable instance directly from a file, which is not directly possible with cf.read (where you have to go through a FieldList to get a Field). This functionality is probably not very useful in practice, since the user is required to gather them in a DataSet to be used in the Component. As a side effect, the inheritence from cf.Field was not playing nicely, especially when trying to use the subspace method on it.

As a result, this commit removes the Variable class, and update the DataSet accordingly, to expect cf.Field instances in place of Variable instances. Moreover, the squeeze functionality is embedded in DataSet directly (used to be a functionality of Variable before), through the argument squeeze of cf.read).

This commit also adds a new keyword argument 'select' for the methods of DataSet, which is also directly given to cf.read. This allows for the selection of specific variables in the data file.

Finally, a new catch-all kwargs is also added to these methods to be given to cf.read, given that cf.read has an extended list of options, they are now accessible for an advanced used this way.