adds a subspace on dataset in _Component
The time series of data available in file may be longer than what is required by the timedomain of the component, so this commit adds a deselection of unwanted time steps (using cf.Field.subspace) before checking the compatibility between the data and the timedomain.

This functionality should also become useful for spinup, where the spinup and simulation periods are likely to be in the same file.