offers possibility to choose the units/calendar when creating TimeDomain with class methods
For this instantiation of a TimeDomain from a datetime sequence or from start/end/step, the units was set to default, and the calendar set to default unless cftime.datetime are given (because they have a calendar attribute). Now, one can optionally choose units and calendar, if not given it falls back to behaviour described above.