adds spin-up functionality
In order to add a spin up functionality, the simulate method of Model is broken down into _initialise, _run, and _finalise so that the new spin_up method can benefit from the two first ones.

New TimeDomains needed to be created for each of the three Components (at the Component level, where timedelta is easily accessible). The method _check_timedomain_compatibilities will be invoked on them as well, and the Clock will be generated from them.

In order to save the user from generating yet another TimeDomain, only the start and end of the spin up run is required, and an optional cycles argument is provided if more than one spin up run is required.

Since the States are owned by the Component, the spin up only needs to 'run' the Component for the required period, and as many times as required, the values of the states are preserved between the cycles and between spin up and simulate (i.e. not reset to zero).

The unittests for model is also modified to add a spin up call before the call to simulate in order to check this new functionality.