removes inheritence of cf.Field for TimeDomain/SpaceDomain
Because the constructors for TimeDomain/SpaceDomain were very different from the constructor of cf.Field, not many methods of Field could be used with TimeDomain/SpaceDomain instances.

Instead, a new attribute f is given to TimeDomain/SpaceDomain to keep the features previously developed, namely to compare these domains with the domains of the data, and to compare the domains cross the components.