adds __repr__ special method for TimeDomain
The method allows to explore the properties of the TimeDomain once initialised. Namely, it should inferred properties such as bounds and timedelta, and also potentially converted properties (e.g. timestamps to datetimes, datetimes in a different calendar, etc.).

Doctrings for contructor methods were also updated to make use of the method, which will be useful when using `doctest`.