splits Grid class into subclasses LatLonGrid and RotatedLatLonGrid
The Grid class featured a lot of tests to be able to deal with both a "true" lat-lon grid, or a grid with a rotated pole. The new implement of Grid does only what is common to both LatLonGrid and RotatedLatLonGrid. The new API for space exposes only these two subclasses. Grid and SpaceDomain are not meant to be used directly.

Also took the opportunity to add a __repr__ method to Grid in order to give an overview of the dimensions and shape of the space domain.

Docstrings were also written for the init of the two subclasses, and alongside the new __repr__ method, they are used for unit tests using doctest.