replaces shape_ attr and makes it a property
The shape_ attribute was set at initialisation which means if the spacedomain axes were updated, the shape_ would become wrong. Now it becomes a property of SpaceDomain.

Also, since the cf.Field is now an attribute of SpaceDomain (not its superclass anymore), there is no risk to name shadow the shape attribute of the field construct, so the trailing underscore is removed.