fixes bug in test model with empty Model dict
The dictionary which was meant to store the Models once initialised was defined in the 'setUp' method. Unfortunately, the doc of unittest says thet the setUp emthod "is called immediately before calling the test method". So the dictionary was reinitialised before each new test method of the class TestCase.

Effectively, this means that the spin_up tests and the simulate tests were not doing anything, because looping over an empty dictionary.

This is now fixed by bringing the said dictionary at the global class level. it won't be reset between tests anymore.