adds a constructor method for regular grid
This method allows to generate a simple regular grid (i.e. LatLonGrid) from the extent and the resolution of its longitude and latitude coordinates.

This method will also be useful to reload a Grid instance from a config file by keeping only the extents and resolutions of the Grid rather than the whole arrays.

The inner functionality is stored in a function in order to allow for a RotatedLatLonGrid to be regenerated as well, however no "public" constructor is given for this class, as the concepts of northwards, eastwards etc. are not straighforward, so it is only given a "private" constructor that will be useful when regenerating from a config file.