adds x/y/z name and units class attributes for Grid
There was a lot of overlap between the two subclasses of Grid's constructor methods from_field. All of this was mostly due to the difference in their standard_name, nothing else. So now Grid and its subclasses have class attributes for Z/Y/X names and their supported units as a list (first item being the preferred option).

Most of the work of the from_field methods is now part of Grid directly to avoid the redundance.

Also adds missing checks for units of altitude.