makes grid/dimension generator functions as "private" class methods of Grid
Since they only make sense for a Grid, it was clumsy to have these functionalities as module functions. They are now Grid class methods, and as such, are accessible to both "public" constructor class of LatLonGrid and "private" constructor class of RotatedLatLonGrid.

The grid generator now generates a dictionary should keys are tailored to the given Grid subclass, and there is no remaping required in the "private" constructor of RotatedLatLonGrid anymore.

The different options to set the YX and Z locations are now given as "private" class attributes (_*_map_loc) rather than hidden in these functions.