adds settings for numerical precision parameters
When numpy.allclose, numpy.isclose are used, the parameters atol and rtol are required for the tolerance for the equality check. So these (ATOL, RTOL) are defined in the settings module. These are functions (like in cf-python).

When numpy.around is used, the number of decimals is required. This is defined in the settings module too (DECR). This is a function as well that allows to change its value.