updates check on altitude in SpaceDomain
For the Z axis, altitude could be used as the standard_name or the computed_standard_name, and the data could have also another property long_name, which would make the .equals method return False, even if everything else matches. So now  it uses a regular expression to find the construct with 'altitude' as an identity. Regular expressions are used for X and Y axes too, just to be safe (usually lat/lon should be the standard_name). Note, the axis Z, Y, X could have been used directly, but they are not compulsary in CF convention, so could not be a fire-proof approach).

It also does not allow for a dataset to have a Z axis if the spacedomain does not have one, which was allowed before. It will prevent some surprises in the use of the data in the component, because beyond the first time dimension T in the data, the ellipsis is used to return the dataset, so it could have returned the altitude axis even if not defined in the spacedomain.

The dummy datasets needed to be updated to include a Z axis. The test_space also needed to be given altitude/altitude_bounds to match the dummy data.