removes squeeze in DataSet and removes direct assignment of Field to DataSet
Squeeze was set as default to True, which was eliminated the dimensions of length one in the Field (typically altitude). Whether a Field has been squeezed, and which fields (and axes) have been squeezed in a DataSet is not straightforward either. This would limit the potential to regenerate a DataSet from dump after crashing. So 'squeeze' option is removed. It is expected that the dataset should be pre-processed if the altitude is to be ignored.

For altitude, the optional argument ignore_z for is_space_equal can be used in the future if needed (in replacement for squeeze).

This means that the init method should not take fields directly anymore, because any preprocessing of the data done at runtime would not be safely preserved. So the only allowed way to populate a DataSet is through files directly. As a result, the from_file classmethod now replaces the init method.