refines check on extent/resolution compatibility
Due to floating point errors, the modulus of end-start  and resolution could be either 0 or resolution for a extent/resolution pair that is genuinely compatible.

So need to add this second case (where the modulus gives a result very close to the value of the resolution, in this case, just like a result very close to 0, it is considered as compatible. The floor division will return an erronous quotient too (i.e. missing one "resolution"), so the dimension size needs to be increased by one (i.e. one resolution size).