makes copies of cfg before popping
If the same spacedomain is used in a config file (i.e. using & * in yaml), if the key 'class' is deleted in the dictionary, it will not be accessible for the next time spacedomian needs to be used for another component. So now it makes a copy of the dict before popping the given key.