fixes bug in timedomain constructor method start_end_step
The sequence of datetimes was generated before converting the datetimes to the (optionally) requested calendar, which would result in problems when converting using cftime.date2num in from_datetime_sequence method, where days may not exist in the requested calendar.

Now, if the calendar is provided, the ymd hms attributes of the datetimes are used to instantiate a cftime datetime of the relevant type before generating the sequence.

Also now add a second example in docstring, because this bug was not caught by the unitest/doctest before. Now it is.