improves spin-up functionality
The spin-up functionality was only able to use a sub-period of the simulation period defined by the component's timedomain.

This was due to the fact that when a component was instantiated, the _check_dataset method was checking the time compatibility between timedomain/dataset and carried out a subspace in time if necessary. So if there was data prior the simulation period in the dataset, it would become unaccessible after the _check_dataset.

The new implementation here extracts the space and time checks in _check_dataset to put them in _check_dataset_space and check_dataset_time, respectively. Now, the _check_dataset and _check_dataset_space still happen during instantiation of the component (since they are common for a spin-up or simulation run), while the check_dataset_time is only carried out in _run of model where it is given the timedomains to check the dataset on, and then does the subspace, that is stored in a new attribute of component _dataset (attribute dataset will store the reference to the pristince cf.Field).

It is now possible to spin-up and simulate on independent periods, whether contiguous, non-contiguous, or even overlapping.

Updates the test in model, to check that spin-up is possible on a period outside of the component timedomain (as long as the period is in the dataset).