initialises fluxes of info through interface at 0 not None
The dummy tests currently don't make use of the fluxes, and so do not use their initial values. But in a real case,  they would, and None could not be used in calculations, so set to default value of zero, but the component's developer will have to provide initial values for these in the future (but it is currently not possible to do so).