removes the last timestamp for TimeDomains
Since the last timestamp was given by the user when creating the TimeDomain instance, it was considered in the TimeDomain 'time' attribute, even though the description said that the last timestamp was the end of the last timestep. This meant that in several places the last timestamp had to be ignored (for comparison with a dataset for example). And in fact, the time bounds were also defined using the last timestamp, which meant that the last bound was going beyond the end of the last timestep (by one timestep length).

This is now chnaged, and the last timestamp given by the user will now be the upper bound of the last time bounds (not its start anymore!).

This involved changes in Clock too to stop the iteration at the last time index now (not the second to last). Also needed to adjust the docstrings to reflect this change.

Note, the _leading and _trailing indices options in the is_time_equal_to are not used anymore, by they are kept in case they come in useful in the future.