makes attributes of component properties
timedomain/spacedomain/dataset/parameters/constants are made properties of component so that checks on them can be run if they are changed after instantiation.

This is required because component could be instantiated from a config file (through model.from_yaml) and then could require some tweaks to make it fit the user's need.