adds Fortran and C versions of Dummy components
The Fortran and C versions implements the initialise-run-finalise methods of the Component classes. DummyFortran and DummyC classes are created, they call these initialise-run-finalise methods.

The Fortran code needs to be compiled with numpy.f2py to get the shared library that Python can call (a makefile is provided).

The C code needs to be compiled with Cython (although other approaches could be used, e.g. building a C extension https://docs.python.org/3/extending/building.html). With the Cython alternative, a pyx file needs to be written in addition to the C file (a makefile is provided).