fixes bug in test for Model
since the check on timedomain compatibility was moved from __init__ method to _run, the instantiation of a Model with Components span different periods would not fail, it would only failed when starting the simulation. So needed to add a call to the simulate method of Model.

Also fixds the name of the method in test_time, starts are different now, not ends anymore.