substitutes timestamp by tag in naming of dump files
For a given identifier, dump files will now be silently overwritten if a new simulation with the same identifier is run. This avoids the building up of tons of dump files, e.g. during testing, or trying to set up the simulation properly.

Information about date and time of creation of dump file is moved to the "description" attribute in the netCDF file.