adds args and kwargs "catchers" for actionless methods of Data/NullComponent
The arguments expected by these methods in the super class Component can be given as positional or keyword arguments depending on the situation. In DataComponent and NullComponent, the arguments were caught sometimes as keyword argument, sometimes as positional argument. It is now harmonised across these actionless methods and *args and **kwargs are added as arguments for all, in order to cover any call case of these methods for the super class Component.